move doTicks to main
